+++ CHANGELOG.md
@@ -4,12 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.27] - 2017-08-02
+## [unreleased]
 ### Added
 - Now constructor return `this`
+- Prefix options
+- Humanize options
+- Server
+    - Process info
+- Browser
+    - User agent info    
 
 ### Changed
-- Now default log folder is "logs" instead of "errors"
+- Server
+    - Now default log folder is "logs" instead of "errors"
+- Browser
+    - Now all logs are saved in "katch" key of localStorage      
 
 ## [0.0.26] - 2017-08-02
 ### Fixed